backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
numpy
opencv-python-headless
pillow
scipy
psycopg2-binary
